Rework Docker image building.
Build a Clang image and use it as the basis for a libgrading variant
rather than adding Clang to libgrading. This allows us to not rebuild
(and re-push) hundreds of megabytes of Clang every time we revise
libgrading. Also, this lets us build the Docker image directly rather
than taking a path through an Ubuntu PPA (which can force us to generate
new public revisions for minor or even test changes).